Write non-'all' playlists too
It's unlikely that I'll ever code something so bad that I'd delete the
"all" playlist. Though, now I've said that...

But, it's not totally unlikely that I'd somehow nuke the other
playlists. To prevent the recovery work being a total pain I've also
chosen to save the video IDs in the output. I suspect - though haven't
confirmed - that there would be a quite easy programmatic way to
reconstruct the playlists given this information alone.